requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
PyJWT
bcrypt==4.0.1
passlib==1.7.4
python-multipart
python-dotenv
pdfplumber
transformers
torch
numpy
scikit-learn
fpdf2
pytest
httpx
datasets
PyMuPDF
